Core/DBLayer: Fix stupid mishap from last commit (i need glasses) - thanks nayd for poking me about it

+++ src/server/game/Accounts/AccountMgr.cpp
@@ -232,7 +232,7 @@ uint32 GetCharactersCount(uint32 accountId)
 {
     // check character count
     PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_SEL_SUM_CHARS);
-    stmt->setUInt64(0, accountId);
+    stmt->setUInt32(0, accountId);
     PreparedQueryResult result = CharacterDatabase.Query(stmt);
 
     return (result) ? (*result)[0].GetUInt64() : 0;
